New "--bumpslopes" exporter option for maketx command
The main purpose of this option is to export additional gradient
informations along bump maps to ease the implementation of bump to
roughness techniques.

The resulting output is a single 6 channels bump map containing the
height, the first and second slopes moment values stored in subsequent
mip-map levels.

Channel data layout:

    channel  quantity             channel name
    -------  --------             ------
      0       h                   "b0_h"
      1       dh/ds               "b1_dhds"
      2       dh/dt               "b2_dhdt"
      3       (dh/ds)^2           "b3_dhds2"
      4       (dh/dt)^2           "b4_dhdt2"
      5       (dh/ds) * (dh/dt)   "b5_dh2dsdt"

Excuse the odd channel names -- because OpenEXR is obliged to re-sort
channels into their alphabetical order, we had to make absolutely sure
that the channels were named in such a way that would not be reordered.

Note that if the input texture is a normal map, then channel[0] = -1.0f
and the corresponding normal can be simply retreived as follows:

    n = normalize( -(dh/ds), -(dh/dt), 1.0 )

The normals are assumed to be defined in the tangent space.

Subtle point: on MIP level 0, channels 3-5 are simple multiples of
channels 1 & 2, but as we filter to get the higher level (lower res)
MIP levels, this is no longer the case, and in fact we depend on this
fact to give us the right standard deviations.

References:

The technique we're using that requires information is based on this
technical report:

Christophe Hery, Michael Kass, and Junhi Ling. "Geometry into Shading."
Technical Memo 14-04, Pixar Animation Studios, 2014.
https://graphics.pixar.com/library/BumpRoughness/
https://graphics.pixar.com/library/BumpRoughness/paper.pdf

And the following references may also be useful in understanding:

Mark Olano and Dan Baker. "LEAN Mapping." Proceedings of I3D 2010.
https://www.csee.umbc.edu/~olano/papers/lean/
https://www.csee.umbc.edu/~olano/papers/lean/lean.pdf

Jonathan Dupuy, et al. "Linear Efficient Antialiased Displacement and
Reflectance (LEADR) Mapping." ACM TOG 32(6), 2013 (Proceedings of
Siggraph Asia 2013).
https://hal.inria.fr/hal-00858220/en
https://hal.inria.fr/hal-00858220v4/document