Speed up TIFF scanlines output (#1833)
TIFF didn't have write_scanlines() of its own, just depending on the
default implementation, which calls write_scanline repeatedly.  So we
give it a proper write_scanlines implementation, which tries to speed
things up by first doing all the conversions and data shuffling on the
whole block (which can be parallelized by our code), before writing the
individual scanlines.

Slightly adjust the chunk size for write_image (it was a little too
conservative).

In parallel_convert_from_float, instead of a hard-coded block size, let
parallel_for_chunked make the choice on its own, which will be better
because it takes into account the number of threads in the pool

These changes, cumulatively, speed up whole-image writes of scanline TIFF
files by about 10%. I was hoping for more than this, but the bottleneck
is the unthreaded code in libtiff that's doing the compression encoding.

This wasn't a problem on the specific images I was testing, but I also
noticed an odditity in the way we did image checkpointing for TIFF
output -- it tries to only checkpoint every 5 seconds (which is longer
than most image writes, so it usually doesn't happen anyway), but the
timer started running when the ImageOutput was created. Seems to me now
that it's better to start the timer when the output opens, not merely
when the ImageOutput is constructed (in case there's a delay between
the two).