oiiotool: autotile off by default. (#1856)
I think it's possible to deadlock when doing certain parallel IBA
functions in combination with autotile. When the deadlock possibility is
fixed, maybe we'll turn it back to on by default, but for now this seems
safer.  Users can always turn it on with the --autotile command line
flag to oiiotool.