Speed up IBA::copy and IBA::crop (#1877)
Both of these were threaded, but used iterators to be fully general.

Much like the recent ImageBuf::get_pixels fix, using
parallel_convert_image (which has been heavily optimized) speeds up the
common situation where the source image is already in local memory and
the ROI being copied is fully contained in the pixel data window.

I'm seeing speedup of copy about 35-45% when copying a 4 channel float
image to either a float buffer, 2.5-3x speedup when a data conversion is
involved (copying a 4 channel float to a 4 channel uint8 buffer), and
2-4x improvement copying 4-chan half to 4-chan half (because the general
code did an expensive half->float->half set of conversions in the
process). Note: the uncertainties are from trying on a few different
machines / processor revisions / number of cores.