New feature: I/O proxies (1931)
ImageOutput and ImageInput supports("ioproxy") will return true for a
format reador or writer that supports the ability to use a
Filesystem::IOProxy object for customized I/O.

For supporting formats, the caller can pass the IOProxy via
a special attribute in the ImageSpec that is passed to
ImageOutput::open() or the configuration ImageSpec passed to
ImageInput::open():

    "oiio:ioproxy" (PTR) is a pointer to a `Filesystem::IOProxy`.

One use for this is to read an image "file" from an in-memory buffer
(using the IOProxy subclass IOMemReader), or to write an image "file" to
an in-memory buffer (using the IOProxy subclass IOVecOutput).

The documentation (in the ImageInput and ImageOutput chapters) and
imagebuf_test.cpp contain examples showing how to do this.

It's also pretty straightforward for both ImageBuf and ImageCache to do
this, since in both cases there is a way to declare the configuration
spec to use when the file is opened.

We implement both reading and writing with I/O proxies for OpenEXR,
and writing with prixies for PNG. More format support will be added
as needed.

Turns out it was harder than I originally thought to absolutely ensure
that a config spec is associated with any opening of an ImageBuf,
incuding when it draws on ImageCache.

Along the way, this necessitated a new variant of ImageInput::create().
So be it.