IC/TS close() and close_all() methods (#1950)
The existing methods invalidate() and invalidate_all() fully invalidate
everything the IC/TS knows about a file (or all files) -- the specs,
the pixels, etc.

But sometimes you just want to close any currently-open handles, but not
necessarily forget about what you know about those images, or any pixel
tiles in the cache. That wasn't possible before, but this patch adds the
functionality as close(filename) and close_all() methods to both IC and
TS.

One interesting use case that this enables is a render-paint loop:

    render();
    ts->close_all();
    ... here, it's ok to modify the texture files ...
    ts->invalidate_all (false);
    render();
    ... repeat

The close_all(), performed after the render, closes any open file
handles, which is especially important on Windows where the IC holding
the file open essentially locks the files. Then, right before the
next render, invalidate_all(false) will fully invalidate any files that
have been modified since they were last opened, but not any whose
modification time has not changed.

Note that the act of closing means that if additional tiles need to be
read from disk, it will need to re-open the file. But as long as
any pixels needed are already in the cache from last time, it may not
need to re-open at all.

To make it easier for apps to detect if the new methods are available,
we define preprocessor symbols OIIO_IMAGECACHE_SUPPORTS_CLOSE and
OIIO_TEXTURESYSTEM_SUPPORTS_CLOSE in imagecache.h and texture.h,
respectively.