Travis: on OSX, don't load libraw from homebrew (#1981)
libraw on homebrew just upgraded to the new 0.19.0. This breaks our raw
tests (builds ok, but seems to not find all the metadata). I'm not sure
if this is a bug in 0.19.0, or if the API changed in some way that we need
to change things on our side. But in either case, it's a more lengthy
investigation than I have time for today, so disabling it on OSX Travis
tests seems like the expedient workaround. Libraw (0.15) is still exercised
on the Linux Travis tests.

As soon as I track down the real problems satisfactorally, I'll restore
the tests.