Print metadata (oiiotool -info -v, etc.) in sorted order.
Alphabetically, but case-insensitive, locale-independent, and with
"un-namespaced" metadata appearing before "namespace-prefixed" metadata.

The main purpose is to canonicalize the order of appearance of metadata
in test reference output. This minimizes the amount of reference output
churn that happens when we change code of image format readers, or upgrade
reader libraries, or add support for new metadata.