Use TravisCI conditions to limit builds (#1983)
We have a growing matrix of tests, and there is development friction
from the fact that every push to a working branch kicks off the whole
matrix, many of which exist to test very specialized things that are
probably not changed by the vast majority of patches.

TravisCI recently started supporting "conditions" that can let you limit
which test cases or stages to run.

I'm proposing that the full test matrix always run when it's:

* a pull request
* a direct push of an updated "master" or "RB" top of tree.
* a nightly cron jobs (which, also, are only for "master" and certain
  "RB" branches)

But in other cases -- that is, ordinary work branch pushes while we
develop -- there are many tests that don't really need to run every
time, and just clog up our Travis queue. So I'm adding the following
exclusions:

- The tests that exist solely to make sure we don't break simd.h (builds
  that exercise AVX2, avx512, and no SIMD) will now only build for
  working branch pushes if the branch name includes the substring
  "simd". (Pro tip: if you are editing simd.h or using it heavily in
  your modified code, name your working branch to include that substring
  to force those tests to run on every push.)

- The tests for gcc8 and/or C++17 (which aren't yet documeted as blessed
  by VFXPlatform) only run for working branch pushes if the branch
  includes the substring "gcc" or "17", respectively.  (Note: since
  gcc6 and C++14 *are* VFXPlatform blessed, those tests are never
  excluded.)

- The test that runs sanitizers, which takes ~4x longer to run than
  other tests, only runs from ordinary working branch pushes if the
  branch name includes the substring "san".

The neat thing is that with the exclusions, an ordinary working branch
push that not have its name indicate that it cares specifically about
simd, san, gcc8, or C++17), only has 5 tests in the remaining matrix,
which is exactly the amount that TravisCI will parallelize. Thus, the
new "common case" will complete its CI in roughly the time it takes to
do just one matrix entry (~10-15 minutes), versus the full test suite
that would have taken 3-4x that.

Remember that this is just for working branch pushes. Once you submit
your work as a PR, the full test matrix will be run.