More robust detection of input bumpmap format. (#2005)
The bumpslopes exporter originally assumed that a 3 channels bumpmap input corresponds to a
normal map. However, depending on the bumpmap creation process, it's not unsual to end up
with grey RGB heightmaps that we incorrectly interpret as normal map.

To prevent this situation we added a safety check to determine if the input map is
monochromatic or not.