ImageCache: add a way to add_tile without a pixel copy (#2012)
The point of this feature is that your app has a buffer of pixels, and
you want the IC or TextureSystem to use that buffer directly, without
copying the texels or owning the memory. It should also not count the
memory against the limits of the cache.

See imagecache_test.cpp, the test I added as text_app_buffer() which
pretty much spells out how to do it from the app point of view.

From an API standpoint, all that changes is an optional argument to
add_tile that lets you disable the copy.

You need to make it a tile of an image in the cache, though, and the
obvious no-drama choice is just to use a "null" ImageInput as a
placeholder.  In the process, I noticed two helpful enhancement I could
make to the null ImageInput better:

1. Make sure that passing a "config" spec results in the right resolution
   and data type -- in fact, there was a bug where the config's res would
   be overwritten by a 1024x1024x4xUINT8 default. Oops. So only overwrite
   those fields if they are not set in the config.

2. What if you want one of these fake inputs but you don't want the name
   of the texture to be "blah.null"? So I added it so that if the config
   has metadata hint "null:force" set to nonzero, it will always succeed
   in creating a null ImageInput.