raw DSLR image read improvements to color space identification. (#2016)
We were previously reporting "oiio:ColorSpace" for raw images always as
"linear", but that's not really how the underlying libraw operates.  It
converts to one of several color spaces as part of the debayering
process (by default, sRGB), so make sure to report back the correct one,
which should be the one you ask for. Also, no need to set the
"raw:ColorSpace" attribute, that doesn't really tell you anything
useful (it just repeats back the hint you asked for, but that's the same
as the "oiio:ColorSpace").