Fix a double free of imagecache and add a replace mode for add_file() (#2021)
* Fix crash deleting not owned imagecache

Now that we allow providing an external ImageCache to the texturesys,
destroy handles its destroying and it shouldn't be done in the
destructor.

* Override existing file entries in add_file

Instead of assuming the present file is equivalent, nuke it and reset it
with the new parameters. This fixes a problem adding files after
somebody has asked for a handle. Also allow non mip-mapped files when
coming from user buffers.