DPX: catch write errors (#2071)
It was reported that if you filled a disk while writing a DPX file, no
errors would be reported, just leaving a corrupted file. (Admittedly,
filling disk volumes is not something I had thoroughly tested.)

The problem turned out to be threefold:

First, inside the DPX library (which was an imported project from
Patrick Palmer, which to the best of my knowledge is no longer actively
maintained outside of OIIO), the underlying OutStream class was used
incorrectly throughout its own library -- in particular, the Write()
method returned the number of bytes correctly written, but the return
results were almost always used incorrectly by merely checking != 0 (as
if it was a bool return for success, when in fact it should check for
any value != the number of bytes being requested to write).

To fix, I added a WriteCheck() method, which returns a simple bool
success/fail, and fixed the call points to use this function. I didn't
replace the original Write() method because there really was one place
where it expected the number of bytes written to be returned.

Second, in our DPXOutput::write_buffer() method, we weren't checking
the return results of the underlying call to m_dpx.WriteElement().
Now we do, and in case of failure try to check `strerror(errno)` in
the hopes of finding a useful system error (and in fact, in this case
of filling the disk, we do!).

Finally, in oiiotool's output_file() function, while we did check the
return codes for the call to out->write_image(), we neglected to check
the return code for out->close(), figuring that nothing could go wrong.
But, oho, the particular workings of our DPX writer are that the
write_scanline calls just batch output, and all the real I/O happens on
the call to close(), so the exact place where the error would occur was
the one place we didn't check. Fun. For most file format types, real
I/O happens in the individual write calls, so this was particular to DPX,
in addition to its other internal bugs described above.

How did I test? On my Mac I created a disk image with the smallest possible
space (around 8MB, I think), then tried to write a big file, like this:

    oiiotool -pattern noise 8000x8000 3 -d uint16 -o /Volumes/SmallTest/out.dpx

which now results in the following error:

    oiiotool ERROR: -o : DPX write failed (No space left on device)

which is about as good as we can hope for.

I also double checked what happens when we fill the disk while writing
tiff, exr, jpg, and they all correctly reported a write error.

So case closed, and thanks to Pavel Jurkas for reporting the DPX error
when the disk volume is full.