Fix instances of possibly unsafe string copying (#2102)
Use of clang-tidy found several instances of "unsafe" string copy or concatenation -- using functions that don't take a buffer length.

In truth, we humans know that these were probably fine in their context, but it's true that those unsafe string functions should be avoided and it's desirable to get clean static analysis on all the security-related tests.

It's worth noting that these changes in rgba.{h,cpp} diverge us from the original sources (they were imported from code by Greg Ward and Bruce Walter). I usually prefer that when we import entire modules of code from elsewhere, we don't make any changes at all, or even reformat to our conventions, so that we can easily integrate future improvements to the upstream sources. But after mulling this over, I think that in this case, it's not actively maintained code, and in 10 years we have never had to integrate fixes. So let's just consider these fully absorbed and able to change at will for our own purposes.