Fix broken TIFF with JPEG compression (#2117)
JPEG compression was one of the cases where we fell back on the
TIFFRGBAImage interface of libtiff. This worked for some JPEG images,
but crashed on others, inside the TIFF library. Never did figure out
why.

But do we really need to to use the RGBAImage interface?  Turns out, no!
At least, not if it's still a spectral RGB image.

So just using the regular interface for ordinary RGB JPEG-compressed
TIFF images is fine, fixes the problem.

Also improved an error message while I was in there.